Refactoring `QutipEmulator` class (#602)
* Adding Hamiltonian class

* fixing type

* Modifying docstring

* Simplifying refactoring

* Simplifying refactoring

* Fix test, simplify attributes in QutipEmulator

* Restore _eval_times_instruction, def device in ham

* delete get_attr

* using NoiseModel for config in Hamiltonian

* Fixing doc, tests

* Move build_operator, add_config to simulation

* Fix typing
